add some service layer tests

+++ src/main/java/by/kirill/service/impl/StatusServiceImpl.java
@@ -1,5 +1,8 @@
 package by.kirill.service.impl;
 
+import by.kirill.controller.handler.exceptions.StatusNotFoundException;
+import by.kirill.controller.handler.exceptions.StatusNotUniqueException;
+import by.kirill.controller.handler.exceptions.StatusNotUpdatebleException;
 import by.kirill.dao.StatusRepository;
 import by.kirill.entity.Status;
 import by.kirill.service.api.StatusService;
@@ -14,8 +17,12 @@ import java.util.Optional;
 @Transactional
 public class StatusServiceImpl implements StatusService {
 
-    @Autowired
-    StatusRepository statusRepository;
+//    @Autowired
+    private StatusRepository statusRepository;
+
+    public StatusServiceImpl(StatusRepository statusRepository) {
+        this.statusRepository = statusRepository;
+    }
 
     @Override
     public List<Status> findAll() {
@@ -23,7 +30,11 @@ public class StatusServiceImpl implements StatusService {
     }
 
     @Override
-    public void deleteById(Integer id) {
+    public void deleteById(Integer id) throws StatusNotUpdatebleException {
+        Status defStatus = statusRepository.findByName("available").get(0);
+        if (id == defStatus.getId()) {
+            throw new StatusNotUpdatebleException();
+        }
         statusRepository.deleteById(id);
     }
 
@@ -38,7 +49,37 @@ public class StatusServiceImpl implements StatusService {
     }
 
     @Override
-    public Status findByName(String name) {
+    public List<Status> findByName(String name) {
         return statusRepository.findByName(name);
+    }
+
+    @Override
+    public Status create(Status status) throws StatusNotUniqueException {
+        if (status.getId() != null || status.getStatusname().isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+        if (!statusRepository.findByName(status.getStatusname()).isEmpty()) {
+            throw new StatusNotUniqueException();
+        }
+        return statusRepository.save(status);
+    }
+
+    @Override
+    public Status update(Status status, Integer id)
+            throws StatusNotFoundException, StatusNotUpdatebleException, StatusNotUniqueException {
+        status.setId(id);
+        Optional<Status> statusToUpdateOpt = statusRepository.findById(status.getId());
+        Status defStatus = statusRepository.findByName("available").get(0);
+        if (!statusRepository.findByName(status.getStatusname()).isEmpty()) {
+            throw new StatusNotUniqueException();
+        }
+        if (status.getId() == defStatus.getId()) {
+            throw new StatusNotUpdatebleException();
+        }
+        if (statusToUpdateOpt.isPresent()) {
+            return statusRepository.save(status);
+        } else {
+            throw new StatusNotFoundException();
+        }
     }
 }
